create /culture html fiel
I wanted to begin this document with an embed code for our Company's Spotify playlist. The playlist serves as a virtual representation of the musical climate that will permeate throughout the company! Interchangeable.. The Starting Five plus a reserve on the bench.. We will apply a coaching method to the DJ'ing. Phil Jackson Style!! Ahh, now I gotta do a one page report on Phil Jackson.. I always got sum to do!